Deserialize FormIndex when retrieving from app preferences
