+++ app/src/org/commcare/tasks/FormLoaderTask.java
@@ -21,6 +21,7 @@ import org.commcare.utils.FileUtil;
 import org.commcare.utils.GlobalConstants;
 import org.javarosa.core.io.StreamsUtil;
 import org.javarosa.core.model.FormDef;
+import org.javarosa.core.model.FormIndex;
 import org.javarosa.core.model.instance.InstanceInitializationFactory;
 import org.javarosa.core.model.instance.TreeElement;
 import org.javarosa.core.model.instance.TreeReference;
@@ -62,7 +63,7 @@ public abstract class FormLoaderTask<R> extends CommCareTask<Integer, String, Fo
     private final SecretKeySpec mSymetricKey;
     private final boolean mReadOnly;
     private final boolean recordEntrySession;
-    private final String serializedLastFormIndex;
+    private final FormIndex lastFormIndex;
 
     private EvaluationTraceReporter traceReporterForFullForm;
     private final boolean profilingEnabledForFormLoad = false;
@@ -75,7 +76,7 @@ public abstract class FormLoaderTask<R> extends CommCareTask<Integer, String, Fo
     public static final int FORM_LOADER_TASK_ID = 16;
 
     public FormLoaderTask(SecretKeySpec symetricKey, boolean readOnly,
-                          boolean recordEntrySession, String formRecordPath, R activity, String serializedFormIndex) {
+                          boolean recordEntrySession, String formRecordPath, R activity, FormIndex lastFormIndex) {
         this.mSymetricKey = symetricKey;
         this.mReadOnly = readOnly;
         this.activity = activity;
@@ -83,7 +84,7 @@ public abstract class FormLoaderTask<R> extends CommCareTask<Integer, String, Fo
         this.recordEntrySession = recordEntrySession;
         this.formRecordPath = formRecordPath;
         TAG = FormLoaderTask.class.getSimpleName();
-        this.serializedLastFormIndex = serializedFormIndex;
+        this.lastFormIndex = lastFormIndex;
     }
 
     /**
@@ -137,7 +138,7 @@ public abstract class FormLoaderTask<R> extends CommCareTask<Integer, String, Fo
 
         setupFormMedia(formDefRecord.getMediaPath());
 
-        AndroidFormController formController = new AndroidFormController(fec, mReadOnly, serializedLastFormIndex);
+        AndroidFormController formController = new AndroidFormController(fec, mReadOnly, lastFormIndex);
 
         data = new FECWrapper(formController);
         return data;
